Implemented multithreeading + async
Now there is 20 threads available to process order
